Refactored backend code by breaking script.js into modules
