+++ index.html
@@ -22,7 +22,7 @@
     />
     <!-- Scripts -->
     <!-- <script defer src="scripts/script.js"></script> -->
-    <script defer src="scripts/prom.js"></script>
+    <script defer type="module" src="src/index.js"></script>
 
     <!-- Styles -->
     <link rel="stylesheet" href="dist/output.css" />
