+++ src/audio.js
@@ -1,0 +1,20 @@
+// ==================================================== //
+// ==== All audio fields are obtained in this file === //
+// ================================================= //
+
+
+// ========================= //
+// ==== Get All Audios ==== //
+// ======================= //
+const letsPlayAudio = document.getElementById("lets-play-audio");
+const selectedAnswerAudio = document.getElementById("selected-answer-audio");
+const correctAnswerAudio = document.getElementById("correct-answer-audio");
+const wrongAnswerAudio = document.getElementById("wrong-answer-audio");
+
+
+export {
+  letsPlayAudio,
+  selectedAnswerAudio,
+  correctAnswerAudio,
+  wrongAnswerAudio
+}
