+++ src/button.js
@@ -1,0 +1,19 @@
+// ==================================================== //
+// ==== All the buttons are obtained in this file ==== //
+// ================================================= //
+
+
+// ===================================================== //
+// ==== Get [Let's Play] Button(Start Game Button) ==== //
+// =================================================== //
+const letsPlayBtn = document.getElementById("lets-play");
+
+// ========================================= //
+// ==== Get Audio/Sound Control Button ==== //
+// ======================================= //
+const audioBtnControl = document.querySelector(".audio-control");
+
+export {
+  letsPlayBtn,
+  audioBtnControl
+}
