+++ src/page.js
@@ -1,0 +1,19 @@
+// ====================================================== //
+// ==== This file contains the game state functions  === //
+// ==================================================== //
+
+
+// ====================================== //
+// ==== Get Welcome Page/Section ======= //
+// ==================================== //
+const welcomeSection = document.querySelector(".welcome-container");
+
+// ======================================= //
+// ==== Get Question Page/Section ======= //
+// ===================================== //
+const questionSection = document.querySelector(".question-container");
+
+export {
+  welcomeSection,
+  questionSection
+}
